frontend/src/pages/api/requirements.txt
requests
beautifulsoup4
openai
newsapi-python
